change to parse as float

+++ gridstatus/miso.py
@@ -68,7 +68,9 @@ class MISO(ISOBase):
 
         mix = {}
         for fuel in r["Fuel"]["Type"]:
-            amount = int(fuel["ACT"])
+            # try as int, then float
+            amount = float(fuel["ACT"])
+
             if amount == -1:
                 amount = 0
             mix[fuel["CATEGORY"]] = amount
